move more properties over to BaseUiTest
